+++ gatsby-config.js
@@ -52,6 +52,14 @@ module.exports = {
       },
     },
     {
+      resolve: `gatsby-plugin-remote-images`,
+      options: {
+        nodeType: 'googleSheetListRow',
+        imagePath: 'profileimage',
+        name: 'localProfileImage',
+      },
+    },
+    {
       resolve: 'gatsby-plugin-emotion',
       options: {
         autoLabel: process.env.NODE_ENV !== 'production',